# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: Vasto Supermarket API
  description: API documentation for the Vasto Supermarket backend
  version: 1.0.0
  servers:
  # - url: 'https://api.vasto.supermarket/dev'  # Development URL
  #   description: Development server (for debugging and testing)
  # - url: 'https://api.vasto.supermarket/prod'  # Production URL
  #   description: Production server (for live environment)
paths:
...
  /api/v1/category:
    post:
      tags:
        - Production
      summary: Add a new category
      description: |
        This endpoint allows the creation of a new category. The request body should contain the category's name, description, and an image file.
        In development, error responses will provide more verbose details.
      operationId: addCategory
      requestBody:
        description: Category details for creation
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the category
                  example: "Electronics"
                description:
                  type: string
                  description: A description of the category
                  example: "Devices and gadgets for everyday use."
                image:
                  type: string
                  format: binary
                  description: The image associated with the category
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Category created successfully"
        '400':
          description: Bad request due to missing or invalid data (e.g., missing image or category already exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Image is required"
        '409':
          description: Category already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Category already exists"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Internal server error"

    get:
      tags:
        - Production
      summary: Get all categories
      description: Fetches all categories available in the system.
      operationId: getAllCategories
      responses:
        '200':
          description: Successfully fetched categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  categories:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: The unique identifier of the category
                          example: "60b6b0e8f3a5f127e45f5b3b"
                        name:
                          type: string
                          description: The name of the category
                          example: "Electronics"
                        description:
                          type: string
                          description: The description of the category
                          example: "Devices and gadgets for everyday use."
        '404':
          description: No categories found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "No categories found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Internal server error"
  /api/v1/category/{id}:
    patch:
      tags:
        - Production
      summary: Update an existing category
      description: |
        This endpoint allows updating an existing category's name, description, and image file.
        In development, error responses will provide more verbose details.
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the category to update.
          schema:
            type: string
      requestBody:
        description: The updated category details
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The new name of the category
                  example: "Electronics"
                description:
                  type: string
                  description: A new description for the category
                  example: "Devices and gadgets for everyday use."
                image:
                  type: string
                  format: binary
                  description: An image file associated with the category
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Category updated successfully"
        '400':
          description: Bad request (invalid data or missing parameters)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid data provided"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Internal server error"

    delete:
      tags:
        - Production
      summary: Delete an existing category
      description: |
        This endpoint allows deleting a category by its ID.
        In development, error responses will provide more verbose details.
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the category to delete.
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Category deleted successfully"
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Category not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Internal server error"
...